BigDecimal changed to double

+++ src/carservice/Invoice.java
@@ -1,23 +1,22 @@
 package carservice;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 // Invoice: Represents an invoice for a customer with total cost and payment status
 public final class Invoice extends CarService {
-    private BigDecimal totalCost;
+    private double totalCost;
     private String paymentStatus;
 
-    public Invoice(BigDecimal totalCost, String paymentStatus) {
+    public Invoice(double totalCost, String paymentStatus) {
         this.totalCost = totalCost;
         this.paymentStatus = paymentStatus;
     }
 
-    public BigDecimal getTotalCost() {
+    public double getTotalCost() {
         return totalCost;
     }
 
-    public void setTotalCost(BigDecimal totalCost) {
+    public void setTotalCost(double totalCost) {
         this.totalCost = totalCost;
     }
